Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 
 int main() {
-    float n;
+    double n;
     printf("");
-    scanf("%f", &n);
-    printf("You entered: %.4f\n", n);
+    scanf("%lf", &n);
+    printf("You entered: %.4lf\n", n);
     return 0;
 }
